+++ Sources/VCore/Services and Managers/Modal Presenter/Link/ModalPresenterLinkViewModifier.swift
@@ -192,6 +192,7 @@ extension View {
 }
 
 private struct SomeModal<Content>: View where Content: View {
+    @Environment(\.modalPresenterInterfaceOrientation) private var interfaceOrientation: PlatformInterfaceOrientation
     @Environment(\.modalPresenterContainerSize) private var containerSize: CGSize
     @Environment(\.modalPresenterSafeAreaInsets) private var safeAreaInsets: EdgeInsets